Completely redesign project, with AST and Lexer
I have completely redesigned this project. I've got a grammar I'm
working on that is EBNF, which handles simple expressions. The
Lexer provides tokens, which now have great test cases! A lot of
work to do still on the Parser class, which will need some extensive
thought. Once the parser is done, it's on to the calculus functions.